[winpr,stream] fix c++ warings for Stream_Write_UINT64

+++ winpr/include/winpr/stream.h
@@ -367,14 +367,14 @@ extern "C"
 	{
 		WINPR_ASSERT(_s);
 		WINPR_ASSERT(Stream_GetRemainingCapacity(_s) >= 8);
-		*_s->pointer++ = (UINT64)(_v)&0xFF;
-		*_s->pointer++ = ((UINT64)(_v) >> 8) & 0xFF;
-		*_s->pointer++ = ((UINT64)(_v) >> 16) & 0xFF;
-		*_s->pointer++ = ((UINT64)(_v) >> 24) & 0xFF;
-		*_s->pointer++ = ((UINT64)(_v) >> 32) & 0xFF;
-		*_s->pointer++ = ((UINT64)(_v) >> 40) & 0xFF;
-		*_s->pointer++ = ((UINT64)(_v) >> 48) & 0xFF;
-		*_s->pointer++ = ((UINT64)(_v) >> 56) & 0xFF;
+		*_s->pointer++ = (_v)&0xFF;
+		*_s->pointer++ = (_v >> 8) & 0xFF;
+		*_s->pointer++ = (_v >> 16) & 0xFF;
+		*_s->pointer++ = (_v >> 24) & 0xFF;
+		*_s->pointer++ = (_v >> 32) & 0xFF;
+		*_s->pointer++ = (_v >> 40) & 0xFF;
+		*_s->pointer++ = (_v >> 48) & 0xFF;
+		*_s->pointer++ = (_v >> 56) & 0xFF;
 	}
 	static INLINE void Stream_Write(wStream* _s, const void* _b, size_t _n)
 	{
